--- keyboards/pica40/config.h
@@ -1,6 +1,0 @@
-// Copyright 2022 zzeneg (@zzeneg)
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-#include "config_common.h"
