+++ keyboards/pica40/keymaps/zzeneg/config.h
@@ -1,0 +1,12 @@
+// Copyright 2022 zzeneg (@zzeneg)
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+
+#define IGNORE_MOD_TAP_INTERRUPT
+#define TAPPING_FORCE_HOLD
+#define TAPPING_FORCE_HOLD_PER_KEY
+#define TAPPING_TERM 150
+#define TAPPING_TERM_PER_KEY
+
+#define BOTH_SHIFTS_TURNS_ON_CAPS_WORD
