+++ keyboards/pica40/keymaps/zzeneg/rules.mk
@@ -1,0 +1,1 @@
+CAPS_WORD_ENABLE = yes
