+++ keyboards/pica40/rev2/rev2.c
@@ -153,3 +153,13 @@ void housekeeping_task_kb(void) {
 
     housekeeping_task_user();
 }
+
+#ifdef ENCODER_ENABLE
+bool encoder_update_kb(uint8_t index, bool clockwise) {
+    if (!encoder_update_user(index, clockwise)) return false;
+
+    tap_code(clockwise ? KC_VOLU : KC_VOLD);
+
+    return false;
+}
+#endif // ENCODER_ENABLE
